try using startNavigationSpan
