+++ static/app/components/devtoolbar/hooks/useCurrentTransactionName.test.tsx
@@ -1,4 +1,4 @@
-import {toSearchTerm} from 'sentry/components/devtoolbar/hooks/useCurrentTransactionName';
+import toSearchTerm from 'sentry/components/devtoolbar/hooks/useCurrentTransactionName';
 
 describe('getSearchTerm', () => {
   it.each([
